Removed home.js & implemented song count and clear songs
Home.js is removed due to redundancy and song counter && clear songs is implemented
