+++ src/Components/MainContent.js
@@ -3,8 +3,8 @@ const MainContent = () => {
 
     return (
         <main>
-            <h1>Your Music Playlist <span><ion-icon name="radio"></ion-icon></span></h1>
-        
+            <h1>Your Music Playlist</h1>
+            
         </main>
         
     );
